Move member/request list components out of "forms" directory
