+++ js/index.js
@@ -20,9 +20,9 @@ import selector from './components/selector'
 import BudgetChart from './components/charts/budget_chart'
 import SpendTable from './components/tables/spend_table'
 import CcpoApproval from './components/forms/ccpo_approval'
-import MembersList from './components/forms/members_list'
+import MembersList from './components/members_list'
 import LocalDatetime from './components/local_datetime'
-import RequestsList from './components/forms/requests_list'
+import RequestsList from './components/requests_list'
 
 Vue.config.productionTip = false
